# [performance] faster Tree expand - eclipse.platform.swt#901