--- a/bundles/org.eclipse.jface/src/org/eclipse/jface/viewers/TreeViewer.java
+++ b/bundles/org.eclipse.jface/src/org/eclipse/jface/viewers/TreeViewer.java
@@ -1150,8 +1150,9 @@ public class TreeViewer extends AbstractTreeViewer {
 			item.dispose();
 
 			// create children on parent
-			for (Object element : children) {
-				createTreeItem(parent, element, -1);
+			// For performance insert every item at index 0 (in reverse order):
+			for (int i = children.length - 1; i >= 0; i--) {
+				createTreeItem(parent, children[i], 0);
 			}
 
 			// reset the selection. client's selection listener should not be triggered.