move setExpanded() after expanding the children

https://github.com/eclipse-platform/eclipse.platform.swt/issues/901