@laeubi 2023-11-27
@SyntevoAlex should this happen for all platforms or only for Windows?