@akurtakov 2023-11-27
Jface should be ws agnostic.