@laeubi 2023-11-27
> Jface should be ws agnostic.

That's true but still we can have platform switches in it if required.