@SyntevoAlex 2023-11-28
I suggest adding a code comment with explanation.

> should this happen for all platforms or only for Windows?

Doing it on all platforms won't hurt, I think. But I didn't test it myself. You could play with my `Bug575787_Tree_SetItemCount_Perf` snippet (already in SWT)