@jukzi 2023-11-28
> I suggest adding a code comment with explanation.

i added comments and reverse order