@basilevs 2023-11-28
## GTK
No regressions
### Before change
﻿210 ms to process 10000 items
202 ms to process 100000 items
The time to select scales as O(n^-0.02) as the size of the model grows from 10000 to 100000
1160 ms to process 10000 items
83267 ms to process 100000 items
The time to reveal scales as O(n^1.86) as the size of the model grows from 10000 to 100000
### After change
196 ms to process 10000 items
399 ms to process 100000 items
The time to select scales as O(n^0.31) as the size of the model grows from 10000 to 100000
1260 ms to process 10000 items
86350 ms to process 100000 items
The time to reveal scales as O(n^1.84) as the size of the model grows from 10000 to 100000
## MacOS
Slowdown?
### Before the change
53 ms to process 10000 items
430 ms to process 100000 items
The time to select scales as O(n^0.90) as the size of the model grows from 10000 to 100000
38 ms to process 10000 items
381 ms to process 100000 items
The time to reveal scales as O(n^1.00) as the size of the model grows from 10000 to 100000
### After the change
110 ms to process 10000 items
1005 ms to process 100000 items
The time to select scales as O(n^0.96) as the size of the model grows from 10000 to 100000
91 ms to process 10000 items
944 ms to process 100000 items
The time to reveal scales as O(n^1.01) as the size of the model grows from 10000 to 100000