@basilevs 2023-11-28
### MacOS increased tree size
The change definitely increases a time to reveal an item in a large tree
### Before
413 ms to process 100000 items
4265 ms to process 1000000 items
The time to select scales as O(n^1.01) as the size of the model grows from 100000 to 1000000
389 ms to process 100000 items
4133 ms to process 1000000 items
The time to reveal scales as O(n^1.03) as the size of the model grows from 100000 to 1000000
### After
969 ms to process 100000 items
9748 ms to process 1000000 items
The time to select scales as O(n^1.00) as the size of the model grows from 100000 to 1000000
963 ms to process 100000 items
9959 ms to process 1000000 items
The time to reveal scales as O(n^1.01) as the size of the model grows from 100000 to 1000000