@jukzi 2023-11-28
> The change definitely increases a time to reveal an item in a large tree

@SyntevoAlex any idea?
@basilevs can you please share the testprogram? i want to try it on windows