@basilevs 2023-11-28
@juksi https://github.com/eclipse-platform/eclipse.platform.ui/issues/649

https://github.com/eclipse-platform/eclipse.platform.ui/files/11019153/bug516281.zip