@jukzi 2023-11-28
> @Juksi #649

jukZi :-)
thanks,. can you post measurements if you combine this PR with 649? all platforms, before+after 
This PR contains 3 different optimizations: 

- move setExpanded() after expanding the children
- insert every item at index 0 (in reverse order)
- setRedraw

can you please figure out which of them does cause regression on Mac?