@basilevs 2023-11-28
> can you post measurements if you combine this PR with 649?
#649 is WIP, there is nothing to combine, I've tested on master and on your branch
> all platforms, before+after
I don't have windows env and already covered Linux and MacOS

> This PR contains 3 different optimizations: 
>     * move setExpanded() after expanding the children
>     * insert every item at index 0 (in reverse order)
>     * setRedraw
> can you please figure out which of them does cause regression on Mac?

I've removed new `setRedraw()` lines from `TreeViewer` and got old performance:

```
395 ms to process 100000 items
4172 ms to process 1000000 items
The time to select scales as O(n^1.02) as the size of the model grows from 100000 to 1000000
366 ms to process 100000 items
4070 ms to process 1000000 items
The time to reveal scales as O(n^1.05) as the size of the model grows from 100000 to 1000000
```