@basilevs 2023-11-28
Performance tests do not work, but are necessary. Life is pain.