@jukzi 2023-11-30
on windows:
```
before PR:
3 ms to process 10000 items
6 ms to process 100000 items
The time to select scales as O(n^0,24) as the size of the model grows from 10000 to 100000
1094 ms to process 10000 items
10413 ms to process 100000 items
The time to reveal scales as O(n^0,98) as the size of the model grows from 10000 to 100000

without redraw(false)
3 ms to process 10000 items
6 ms to process 100000 items
The time to select scales as O(n^0,27) as the size of the model grows from 10000 to 100000
995 ms to process 10000 items
9993 ms to process 100000 items
The time to reveal scales as O(n^1,00) as the size of the model grows from 10000 to 100000

with redraw(false)
5 ms to process 10000 items
13 ms to process 100000 items
The time to select scales as O(n^0,38) as the size of the model grows from 10000 to 100000
3105 ms to process 10000 items
30691 ms to process 100000 items
The time to reveal scales as O(n^0,99) as the size of the model grows from 10000 to 100000
```

=> that test can not show that this PR  can make some things faster, but also no regression
=> with this test redraw(false) is not good. probably because the tested view is much less complicated then the Problem View i measured with. So probably it's better to use redraw(false) not in general but may be in Problem View