@basilevs 2023-12-01
@jukzi have you tried using` Tree.setItemCount()` instead of creating items one-by-one? I was under impression, it should dramatically increase performance.