@jukzi 2023-12-01
> @jukzi have you tried using` TreeItem.setItemCount()` instead of creating items one-by-one? IMO, it should dramatically increase performance.

I have not tried that. i don't understand how it should be used. Is that replacement for multiple `new TreeItem(parent, -1)` ?