@basilevs 2023-12-01
> Is that replacement for multiple `new TreeItem(parent, -1)` ?

Yes. The change would be to replace multiple new TreeItem() with setItemCount()
than traverse those with `getItem(index)`, `updateItem()`and `updatePlus()`. Performance gain is not guaranteed and would require testing (as `getItem(int)` may be slow enough to negate any gains).