@jukzi 2023-12-01
We got a Problem: this change keeps failing on linux - and only linux  - 

[org.eclipse.jface.tests.viewers.VirtualLazyTreeViewerTest.testExpandToLevel](https://ci.eclipse.org/platform/job/eclipse.platform.ui/job/PR-1331/8/testReport/junit/org.eclipse.jface.tests.viewers/VirtualLazyTreeViewerTest/testExpandToLevel/)
[org.eclipse.jface.tests.viewers.TreeViewerWithLimitTest.testRemoveItemsAtParent](https://ci.eclipse.org/platform/job/eclipse.platform.ui/job/PR-1331/8/testReport/junit/org.eclipse.jface.tests.viewers/TreeViewerWithLimitTest/testRemoveItemsAtParent/)

The code should be theoretically OS independent. i am on windows. @basilevs can you look why it fails on linux?