@laeubi 2023-12-01
@jukzi first basic step would be to enhance the test with *what* assertions failed and probably adding a message that helps in understand the test failure.