@basilevs 2023-12-01
The test fails to expand second level of the tree:
<img width="562" alt="Screenshot 2023-12-01 at 13 55 19" src="https://github.com/eclipse-platform/eclipse.platform.ui/assets/650857/bf5533b1-69c6-445f-8013-734c5f1dbc18">